benchmarking: move the common functions into uucore
