+++ src/uu/tsort/Cargo.toml
@@ -30,6 +30,7 @@ path = "src/main.rs"
 [dev-dependencies]
 divan = { workspace = true }
 tempfile = { workspace = true }
+uucore = { workspace = true, features = ["benchmark"] }
 
 [[bench]]
 name = "tsort_bench"
