+++ src/uucore/Cargo.toml
@@ -47,6 +47,9 @@ base64-simd = { version = "0.8", optional = true }
 libc = { workspace = true, optional = true }
 os_display = "0.1.3"
 
+# Benchmark dependencies (optional)
+divan = { workspace = true, optional = true }
+
 digest = { workspace = true, optional = true }
 hex = { workspace = true, optional = true }
 memchr = { workspace = true, optional = true }
@@ -169,3 +172,4 @@ wide = []
 tty = []
 time = ["jiff"]
 uptime = ["chrono", "libc", "windows-sys", "utmpx", "utmp-classic"]
+benchmark = ["divan"]
